Fix DesireWarWith Asserts
Recompute DesireWar when
- declaring war
- ending war
- making contact
- turn begins

Corrected computation of DesireWar on load

When a civ is killed the other civs loose contact and all their embassy with that civ. This is also reset when the slot is reused for a new civ, so that there is no embassy and contact with the civ.