Improve debug stack printing and add a meaningful line to crash.txt in Win 64 builds
The addresses in the debug stack have now 16 digits in 64 bit builds and the Windows 64 bit release builds generate now give the function where a crash occurred. Unlike the 32 bit counterpart, which generates a full stack trace, it is just the top function in release, despite that debug builds generate this way the full stack trace. Disabling frame pointer suppression for 64 bit builds did not change anything. So this is not part of this commit.