Add a garrison force match to determine the needed garrison strength
The garrison force match is lower than the defense force match, so that the garrison won't be as big as needed for the actual defense goal. This frees up units that can then be used for offesive ones enough units have been built.

This also allows to add a new goal type to goals.txt, with the garrison force match.