Add sub-goal for goals
This adds subgoals for other goals, for instance the defense goal has now the subgoal garrison. This way the AI can stuff its cities with defense units partically. This has the advantage over garrison that the AI can now move units for city garrison between cities instead of relying on units that come for city siege into these cities and then are removed from the siege, making the siege goals fail.

However, there still seems to be a problem with ccity garison so the subgoal garrison is still not in strategies.txt.